backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.24.1
supabase==2.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langdetect==1.0.9
langid==1.1.5
slowapi==0.1.9
redis==5.0.1
replicate==0.22.0
ibm-watson==7.0.1
